% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-pr.R
\name{create_gh_pull_request}
\alias{create_gh_pull_request}
\title{Create GitHub Pull Request}
\usage{
create_gh_pull_request(
  source_branch = gert::git_branch(),
  target_branch = "main",
  repo_url = gitworkr::get_remote_repo_url(),
  pr_template = NULL
)
}
\arguments{
\item{source_branch}{The name of the branch containing changes to be merged
into the \code{target_branch}. The current branch is used by default.}

\item{target_branch}{The name of the receiving branch, which is typically the
"main" (default) branch.}

\item{repo_url}{The GitHub repository URL where the Pull Request will be
created. By default, \code{gitworkr::get_remote_repo_url()} is used to parse the
URL of the remote "origin".}

\item{pr_template}{The name of a pull request template file located in the
subdirectory ".github/PULL_REQUEST_TEMPLATE/" or
"docs/PULL_REQUEST_TEMPLATE/" of the repo specified in \code{repo_url}.}
}
\description{
Create GitHub Pull Request
}
