% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-ops.R
\name{gen_repo_from_template}
\alias{gen_repo_from_template}
\title{Create new repository from GitHub template repo}
\usage{
gen_repo_from_template(
  repo_owner,
  repo_name,
  repo_descr,
  proj_dir,
  tmplt_owner,
  tmplt_repo
)
}
\arguments{
\item{repo_owner}{github username and owner of the new github repository.}

\item{repo_name}{name of new the github repository being created.}

\item{repo_descr}{description for the new github repository.}

\item{proj_dir}{target path to clone repo and initialize RStudio project.}

\item{tmplt_owner}{username associated with github template repository.}

\item{tmplt_repo}{name of github template repository.}
}
\value{
list containing github API response object \code{gh_response} as well as
the local path to the created RStudio project
}
\description{
Create new repository from GitHub template repo
}
\examples{
\dontrun{
  ## create a new repo from a github template repository
  gen_repo_from_template(
    repo_owner = "",
    repo_name = "",
    repo_descr = "",
    proj_dir = "",
    tmplt_owner = "",
    tmplt_repo = ""
  )
}

}
