% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-ops.R
\name{gen_repo_from_template}
\alias{gen_repo_from_template}
\title{Create new repository from GitHub template repo}
\usage{
gen_repo_from_template(
  repo_owner,
  repo_name,
  repo_descr = "",
  proj_dir,
  tmplt_owner,
  tmplt_repo
)
}
\arguments{
\item{repo_owner}{github username and owner of the new github repository.}

\item{repo_name}{name of new the github repository being created.}

\item{repo_descr}{description for the new github repository.}

\item{proj_dir}{target path for cloning repo and initializing RStudio
project.}

\item{tmplt_owner}{username associated with github template repository.}

\item{tmplt_repo}{name of github template repository.}
}
\value{
list containing github API response object \code{gh_response} as well as
the local path to the new RStudio project
}
\description{
Create a new GitHub repository from a template repo, clone the new repo to
your local machine, create an RStudio project for the cloned repository, and
open the RStudio project in a separate R session.
}
\examples{
\dontrun{
  ## create a new repo from a github template repository
  gen_repo_from_template(
    repo_owner = "ilinsmeier",
    repo_name = "new-github-repo-from-template",
    repo_descr = "This repository was created from a github template repo.",
    proj_dir = "~/projects",
    tmplt_owner = "ilinsmeier",
    tmplt_repo = "bg-labs-r-project-template"
  )
}

}
