% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-ops.R
\name{gen_repo_from_template}
\alias{gen_repo_from_template}
\title{Create new repository from GitHub template repo}
\usage{
gen_repo_from_template(
  repo_owner,
  repo_name,
  repo_descr = "",
  proj_dir,
  tmplt_owner,
  tmplt_repo,
  clean_readme = TRUE
)
}
\arguments{
\item{repo_owner}{github username (string)}

\item{repo_name}{name of the new the github repository being created
(string).}

\item{repo_descr}{a description for the new github repository (string).}

\item{proj_dir}{target path to clone repository and initialize RStudio
project (string).}

\item{tmplt_owner}{username associated with github template repository
(string).}

\item{tmplt_repo}{name of github template repository (string).}

\item{clean_readme}{create "README.md" or replace file contents (if it
already exists) with \code{repo_name} & \code{repo_descr} (logical, default = TRUE).}
}
\value{
list containing github API response object \code{gh_response} as well as
the local path to the new RStudio project
}
\description{
Create a new GitHub repository from a template repo, clone the new repo to
your local machine, create an RStudio project for the cloned repository, and
open the RStudio project in a separate R session.
}
\examples{
\dontrun{
  ## create a new repo from a github template repository
  gen_repo_from_template(
    repo_owner = "ilinsmeier",
    repo_name = "new-github-repo-from-template",
    repo_descr = "This repository was created from a github template repo.",
    proj_dir = "C:/Users/ilinsmeier/projects",
    tmplt_owner = "ilinsmeier",
    tmplt_repo = "bg-labs-r-project-template",
    clean_readme = TRUE
  )
}

}
